# app/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RustAcademy Backend API
  description: >
    RustAcademy is a Stellar-based payment platform enabling payment links,
    recurring subscriptions, username resolution, marketplace trading,
    fiat on/off ramps, and webhook-driven event notifications.

    **Base URL (production):** `https://api.RustAcademy.example.com`
    **Base URL (development):** `http://localhost:3000`

    Interactive Swagger UI is available at `/docs` when the server is running.
  version: 1.0.0
  contact:
    name: RustAcademy API Support
    url: https://github.com/RustAcademy/RustAcademy
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.RustAcademy.example.com
    description: Production
...
security:
  - ApiKeyAuth: []
  - {}
paths:
...
  /status:
    get:
      tags: [health]
      operationId: getPublicStatus
      summary: Public status page
      description: >
        Returns a minimal, cacheable status suitable for a public status page.
        No sensitive operational details are exposed.
        Helps users distinguish client errors vs platform downtime.
      security: []
      responses:
        "200":
          description: Public status information
          headers:
            ETag:
              schema:
                type: string
              description: Cache validator for conditional requests
            Cache-Control:
              schema:
                type: string
              description: Caching directives
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicStatusResponse"
        "304":
          description: Not modified (use cached version)
        "429":
          description: Too many requests (rate limited)

  # ---------------------------------------------------------------------------
  # Usernames
  # ---------------------------------------------------------------------------
...
  /telegram/status/{telegramId}:
    get:
      tags: [Telegram]
      operationId: getTelegramStatus
      summary: Check Telegram account linkage status
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: integer
          description: Telegram user ID
      responses:
        "200":
          description: Linkage status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramLinkStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
  /telegram/verify/{telegramId}:
    post:
      tags: [Telegram]
      operationId: verifyTelegramLink
      summary: Verify Telegram account linkage
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verificationCode]
              properties:
                verificationCode:
                  type: string
      responses:
        "200":
          description: Verification successful
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /telegram/settings/{telegramId}:
    put:
      tags: [Telegram]
      operationId: updateTelegramSettings
      summary: Update Telegram notification settings
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTelegramSettingsRequest"
      responses:
        "200":
          description: Settings updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /telegram/link/{telegramId}:
    delete:
      tags: [Telegram]
      operationId: unlinkTelegram
      summary: Unlink Telegram account
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Account unlinked
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Reconciliation
  # ---------------------------------------------------------------------------
...
  /reconciliation/trigger:
    post:
      tags: [reconciliation]
      operationId: triggerReconciliation
      summary: Manually trigger a reconciliation run
      responses:
        "200":
          description: Reconciliation completed
        "409":
          description: Already running
  /reconciliation/backfill:
    post:
      tags: [reconciliation]
      operationId: startBackfill
      summary: Trigger a backfill job for a ledger range
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackfillConfig"
      responses:
        "200":
          description: Backfill completed
        "409":
          description: Backfill already running
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for higher rate limits

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            request_id:
              type: string
              format: uuid
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            details:
              type: object

    # --- Health ---
...
    PublicStatusComponent:
      type: object
      properties:
        name:
          type: string
          example: horizon
        status:
          type: string
          enum: [operational, degraded, down]
          example: operational
        detail:
          type: string
          example: "Network: testnet"
    PublicStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [operational, degraded, down]
          example: operational
          description: Overall platform status
        network:
          type: string
          example: testnet
          description: Stellar network name
        lastLedger:
          type: integer
          example: 12345678
          description: Last processed ledger sequence
        timestamp:
          type: string
          format: date-time
        components:
          type: array
          items:
            $ref: "#/components/schemas/PublicStatusComponent"
        version:
          type: string
          example: "0.1.0"

    # --- Usernames ---
...
    TelegramLinkStatus:
      type: object
      properties:
        isLinked:
          type: boolean
        isVerified:
          type: boolean
        publicKey:
          type: string
          nullable: true
        enabled:
          type: boolean
          nullable: true
        minAmountXlm:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
    UpdateTelegramSettingsRequest:
      type: object
      properties:
        enabled:
          type: boolean
        minAmountStroops:
          type: string

    # --- Reconciliation ---
    BackfillConfig:
      type: object
      properties:
        startLedger:
          type: integer
        endLedger:
          type: integer

    # --- Job Queue ---
...